+++ CHANGELOG.md
@@ -11,6 +11,14 @@ to make parsing easier with multiple symbols as arguments for calls.
 Useful for storing callbacks.
 * Added user field to Env, so the user of wlambda can provide his own registry for
 storing data from globally added functions.
+* Added error values for error reporting, and \_? and on\_error functions for
+handling error values. \_? returns error values directly upwards, like ? in
+Rust. And on\_error handles an error value in the function passed as first
+argument.
+* Added block and return for non local jumps upwards the stack.
+* Added syntax: \:func_label { ... } for easier returns.
+* Added syntax: \ \<statement\>  for one statement functions.
+* Added panic for fatal error reporting.
 
 0.1.1 (2019-05-23)
 ==================